+++ src/Core.sol
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.23;
+
+/* solhint-disable no-unused-import */
+import { ExtensibleFallbackHandler } from "./core/ExtensibleFallbackHandler.sol";