+++ src/Mocks.sol
@@ -1,0 +1,22 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.23;
+
+/* solhint-disable no-unused-import */
+/*//////////////////////////////////////////////////////////////
+                             Aux
+//////////////////////////////////////////////////////////////*/
+import { MockRegistry } from "module-bases/mocks/MockRegistry.sol";
+import { MockTarget } from "module-bases/mocks/MockTarget.sol";
+
+/*//////////////////////////////////////////////////////////////
+                             Modules
+//////////////////////////////////////////////////////////////*/
+import { MockValidator } from "module-bases/mocks/MockValidator.sol";
+import { MockExecutor } from "module-bases/mocks/MockExecutor.sol";
+import { MockHook } from "module-bases/mocks/MockHook.sol";
+
+/*//////////////////////////////////////////////////////////////
+                             Tokens
+//////////////////////////////////////////////////////////////*/
+import { MockERC20 } from "forge-std/mocks/MockERC20.sol";
+import { MockERC721 } from "forge-std/mocks/MockERC721.sol";