+++ src/Modules.sol
@@ -1,0 +1,19 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.23;
+
+/* solhint-disable no-unused-import */
+import { ERC7579ValidatorBase } from "module-bases/ERC7579ValidatorBase.sol";
+import { ERC7579ExecutorBase } from "module-bases/ERC7579ExecutorBase.sol";
+import { ERC7579HookBase } from "module-bases/ERC7579HookBase.sol";
+import { ERC7579HookDestruct } from "module-bases/ERC7579HookDestruct.sol";
+import { ERC7579FallbackBase } from "module-bases/ERC7579FallbackBase.sol";
+import { SchedulingBase } from "module-bases/SchedulingBase.sol";
+import {
+    IERC7579Validator,
+    IERC7579Executor,
+    IERC7579Fallback,
+    IERC7579Hook
+} from "./external/ERC7579.sol";
+import { ERC7484RegistryAdapter } from "module-bases/ERC7484RegistryAdapter.sol";
+import { ERC7579ModuleBase } from "module-bases/ERC7579ModuleBase.sol";
+import { TrustedForwarder } from "module-bases/utils/TrustedForwarder.sol";